fix: remove broken `--cut-string` option
If we enable $k_T$ width, it moves the Lund string quarks' event record
numbers, breaking the implementation of `--cut-string`. It turns out
that we do not need this option, since according to Albi, it's only
useful for the analyzing power associated with $G_1^\perp$, but not for
the full asymmetry.